Move random meme function, add meme_message file
The function to select a random meme from the library was moved to
within the file containing the library. The version of the meme
machine that was previously used for sending long messages such as
a welcome has been added, now that it can send an arbitrary message.
It takes an additional command line argument for the file containing
the message to be sent, then reads that file and splits it into
multile 155-character messages, which are sent with a small delay in
between each to ensure they arrive in proper order.